external consumers - re-organize connector providers
