+++ broker/pom.xml
@@ -29,5 +29,6 @@
         <module>core</module>
         <module>plugin</module>
         <module>client</module>
+        <module>connector-amqp</module>
     </modules>
 </project>
