+++ connector/api/src/main/java/org/eclipse/kapua/connector/MessageContext.java
