--- connector/commons/pom.xml
@@ -1,27 +1,0 @@
-<?xml version="1.0"?>
-<!--
-    Copyright (c) 2018 Eurotech and/or its affiliates and others
-
-    All rights reserved. This program and the accompanying materials
-    are made available under the terms of the Eclipse Public License v1.0
-    which accompanies this distribution, and is available at
-    http://www.eclipse.org/legal/epl-v10.html
-
-    Contributors:
-        Eurotech - initial API and implementation
- -->
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-
-    <parent>
-        <groupId>org.eclipse.kapua</groupId>
-        <artifactId>kapua-connector</artifactId>
-        <version>1.1.0-SNAPSHOT</version>
-    </parent>
-
-    <artifactId>kapua-connector-commons</artifactId>
-    <name>${project.artifactId}</name>
-
-    <dependencies>
-    </dependencies>
-</project>
