+++ connector/provider/hono/src/main/java/org/eclipse/kapua/connector/hono/settings/ConnectorHonoSettingKey.java
