+++ connector/provider/kura/src/main/java/org/eclipse/kapua/connector/kura/KuraPosition.java
