+++ connector/provider/logger/src/main/java/org/eclipse/kapua/connector/logger/LoggerTargetJAXBContextProvider.java
