+++ service/datastore/pom.xml
@@ -31,6 +31,7 @@
         <module>client-embedded</module>
         <module>client-rest</module>
         <module>client-transport</module>
+        <module>connector</module>
     </modules>
 
 </project>
